move test into lesson folder
