+++ frontend/src/features/lesson/test/questionCard/AnswerVariant/AnswerVariant.tsx
