+++ frontend/src/features/lesson/test/questionCard/AnswerVariant/questionsData.ts
